simple_story_generator/requirements.txt
langchain_huggingface==0.1.2
python-dotenv==1.0.1
streamlit==1.43.2